+++ src/main/java/com/sora/ryokka/service/EmployeeImagesService.java
@@ -3,6 +3,6 @@ package com.sora.ryokka.service;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface EmployeeImagesService {
-    String uploadEmployeeImage(MultipartFile file, Integer employeeId);
-    void updateEmployeeImage(Integer employeeId, String imageUrl);
+    String uploadEmployeeImage(MultipartFile file, Long employeeId);
+    void updateEmployeeImage(Long employeeId, String imageUrl);
 }
